requirements.txt
grpcio-tools~=1.38.1
matplotlib~=3.4.2
numpy~=1.21.0
scikit-learn~=0.24.2